backend/alert-service/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
protobuf==4.25.1
gtfs-realtime-bindings==1.0.0
APScheduler==3.10.4
redis==5.0.1
pytest==7.4.3
pytest-cov==4.1.0